Split up CreateDefinition
